+++ packages/plugin-packages/mashroom-websocket/src/backend/webapp/mashroom-bootstrap-webapp.js
@@ -1,0 +1,38 @@
+// @flow
+
+import context from '../context';
+import expressApp from './webapp';
+import httpUpgradeHandlerFn from './http_upgrade_handler';
+import WebSocketServer from '../WebSocketServer';
+
+import type {
+    MashroomHttpUpgradeHandler,
+    MashroomWebAppPluginBootstrapFunction
+} from '@mashroom/mashroom/type-definitions';
+
+const bootstrap: MashroomWebAppPluginBootstrapFunction = async (pluginName, pluginConfig, contextHolder) => {
+    const { path, restrictToRoles, pingIntervalSec, maxConnections } = pluginConfig;
+    const pluginContext = contextHolder.getPluginContext();
+
+    const server = new WebSocketServer(pluginContext.loggerFactory);
+    context.server = server;
+    context.restrictToRoles = restrictToRoles;
+    context.basePath = path;
+    context.pingIntervalSec = pingIntervalSec;
+    context.maxConnections = maxConnections;
+
+    const upgradeHandler: MashroomHttpUpgradeHandler = httpUpgradeHandlerFn(pluginContext.loggerFactory);
+
+    const onUnload = () => {
+        // Close all connections when the plugin reloads
+        server.closeAll();
+    };
+
+    return {
+        expressApp,
+        upgradeHandler,
+        onUnload,
+    };
+};
+
+export default bootstrap;